ia3d: A3DImport: read A3DCreatorMaterial with A3DInputStream

+++ ia3d/src/main/java/good/damn/ia3d/models/A3DMAsset.kt
@@ -1,5 +1,5 @@
 package good.damn.ia3d.models
 
 data class A3DMAsset(
-    val version: Int
+    val version: Short
 )
